Move indexer to lib
