+++ lib/bedrock/index.js
@@ -1,4 +1,4 @@
-const buildIndexFromArray = require('./indexer')
+const buildIndexFromArray = require('../indexer')
 
 module.exports = (data) => {
   return {
