+++ lib/indexer.js
